% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gerar_tabela}
\alias{gerar_tabela}
\title{Gerar tabela com países e respectivas capitais}
\usage{
gerar_tabela(caminho)
}
\arguments{
\item{caminho}{Caminho onde será salva a base de dados no seu computador.}
}
\value{
Retorna uma tibble com a tabela com os países, capitais e continentes.
}
\description{
Gerar tabela com países e respectivas capitais
}
