% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gerar_tabela}
\alias{gerar_tabela}
\title{Gera uma tabela com países, capitais e continentes do mundo}
\usage{
gerar_tabela(salvar = FALSE, formato = "csv")
}
\arguments{
\item{salvar}{Decidir se deve ser salvo um arquivo local no computador com a tabela. Valor padrão: \code{FALSE}}

\item{formato}{Formato do arquivo a ser salvo localmente. Por padrão, "csv". Disponíveis}
}
\value{
Retorna uma tibble contendo países, capitais e seus respectivos continentes.
}
\description{
Gera uma tabela com países, capitais e continentes do mundo
}
