% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gerar_tabela}
\alias{gerar_tabela}
\title{Generate table with countries, capitals e continents}
\usage{
gerar_tabela()
}
\value{
Tibble containing countries, capitals and their respective continents.
}
\description{
Generate table with countries, capitals e continents
}
